+++ packages/web/src/common/components/MyMenu/index.tsx
@@ -5,7 +5,7 @@ import styled from "styled-components";
 
 import { useAuth } from "@sparcs-clubs/web/common/providers/AuthContext";
 
-import Button from "../Buttons/Button";
+import Button from "../Button";
 import Icon from "../Icon";
 
 import ProfileList from "./_atomic/ProfileList";
