+++ packages/web/src/features/common-space/frames/CommonSpaceInfoFrame/index.tsx
@@ -4,7 +4,7 @@ import { subSeconds } from "date-fns";
 
 import styled from "styled-components";
 
-import Button from "@sparcs-clubs/web/common/components/Buttons/Button";
+import Button from "@sparcs-clubs/web/common/components/Button";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import StepProcess from "@sparcs-clubs/web/common/components/StepProcess/StepProcess";
