+++ packages/web/src/features/my/activity-certificate/frame/MyCertificateDetailFrame.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { useRouter } from "next/navigation";
 import styled from "styled-components";
 
-import Button from "@sparcs-clubs/web/common/components/Buttons/Button";
+import Button from "@sparcs-clubs/web/common/components/Button";
 import Card from "@sparcs-clubs/web/common/components/Card";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import {
