+++ packages/web/src/features/register-club/components/RulesButton.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import styled from "styled-components";
 
-import Button from "@sparcs-clubs/web/common/components/Buttons/Button";
+import Button from "@sparcs-clubs/web/common/components/Button";
 import Typography from "@sparcs-clubs/web/common/components/Typography";
 
 interface RulesButtonProps {
